requirements.txt
fastapi==0.85.1
pydantic==1.10.2
python-dotenv==0.21.0
python-multipart==0.0.5
requests==2.28.1
slack-sdk==3.19.1
starlette==0.20.4
uvicorn==0.18.3